% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{backfill_dicts}
\alias{backfill_dicts}
\title{Backfill Dicts}
\usage{
backfill_dicts(col)
}
\arguments{
\item{col}{A list of dictionaries, where each dictionary might have missing keys}
}
\value{
A list of dictionaries, where each dictionary contains all possible keys
}
\description{
Fills missing values in a list of dictionaries
}
