% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_all_drugs}
\alias{get_all_drugs}
\title{Get All Drugs}
\usage{
get_all_drugs(api_url = NULL)
}
\arguments{
\item{api_url}{API endpoint for GraphQL request, Default: NULL}
}
\value{
a full list of drugs present in dgidb
}
\description{
Get all drug names present in DGIdb
}
\examples{
get_all_drugs()
}
