% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_all_genes}
\alias{get_all_genes}
\title{Get All Genes}
\usage{
get_all_genes(api_url = NULL)
}
\arguments{
\item{api_url}{API endpoint for GraphQL request, Default: NULL}
}
\value{
a full list of genes present in dgidb
}
\description{
Get all gene names present in DGIdb
}
\examples{
get_all_genes()
}
