% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_categories}
\alias{get_categories}
\title{Get Categories}
\usage{
get_categories(terms, api_url = NULL)
}
\arguments{
\item{terms}{Genes of interest for annotations}

\item{api_url}{API endpoint for GraphQL request, Default: NULL}
}
\value{
category annotation results for genes
}
\description{
Perform a category annotation lookup for genes of interest
}
\examples{
get_categories(c("BRAF", "PDGFRA"))
}
