% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_drug_applications}
\alias{get_drug_applications}
\title{Perform a look up for ANDA/NDA applications for drug or drugs of interest}
\usage{
get_drug_applications(terms, use_processing = TRUE)
}
\arguments{
\item{terms}{drugs of interest}

\item{use_processing}{placeholder}
}
\value{
all ANDA/NDA applications for drugs of interest
}
\description{
Perform a look up for ANDA/NDA applications for drug or drugs of interest
}
\examples{
terms <- "DOVITINIB"
get_drug_applications(terms)
}
