% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_drugs}
\alias{get_drugs}
\title{Get Drugs}
\usage{
get_drugs(terms, immunotherapy = NULL, antineoplastic = NULL, api_url = NULL)
}
\arguments{
\item{terms}{drugs for record lookup}

\item{immunotherapy}{filter option for results that are only immunotherapy, Default: NULL}

\item{antineoplastic}{filter option for results that see antineoplastic use, Default: NULL}

\item{api_url}{API endpoint for GraphQL request, Default: NULL}
}
\value{
drug data
}
\description{
Perform a record look up in DGIdb for a drug of interest
}
\examples{
get_drugs(c("Imatinib"))
}
