% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Get all gene names present in DGIdb}
\usage{
get_gene_list()
}
\value{
full list of all genes
}
\description{
Get all gene names present in DGIdb
}
\examples{
get_gene_list()
}
