% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_genes}
\alias{get_genes}
\title{Get Genes}
\usage{
get_genes(terms, api_url = NULL)
}
\arguments{
\item{terms}{genes for record lookup}

\item{api_url}{API endpoint for GraphQL request, Default: NULL}
}
\value{
gene data
}
\description{
Perform a record look up in DGIdb for genes of interest
}
\examples{
get_genes(c("BRAF", "PDGFRA"))
}
