% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_interactions}
\alias{get_interactions}
\title{Get Interactions}
\usage{
get_interactions(
  terms,
  search = "genes",
  immunotherapy = NULL,
  antineoplastic = NULL,
  source = NULL,
  pmid = NULL,
  interaction_type = NULL,
  approved = NULL,
  api_url = NULL
)
}
\arguments{
\item{terms}{drugs or genes for interaction look up}

\item{search}{interaction search type. valid types are "drugs" or "genes", Default: 'genes'}

\item{immunotherapy}{filter option for results that are used in immunotherapy, Default: NULL}

\item{antineoplastic}{filter option for results that are part of antineoplastic regimens,
Default: NULL}

\item{source}{filter option for specific database of interest, Default: NULL}

\item{pmid}{filter option for specific PMID, Default: NULL}

\item{interaction_type}{filter option for specific interaction types, Default: NULL}

\item{approved}{filter option for approved interactions, Default: NULL}

\item{api_url}{API endpoint for GraphQL request, Default: NULL}
}
\value{
interaction results for terms
}
\description{
Perform an interaction look up for drugs or genes of interest
}
\examples{
get_interactions(c("BRAF", "PDGFRA"))
}
