% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{group_attributes}
\alias{group_attributes}
\title{Group Attributes}
\usage{
group_attributes(row)
}
\arguments{
\item{row}{A list of dictionaries of attributes}
}
\value{
A dictionary which maps each attribute to a list of associated values
}
\description{
Groups attributes by name and stores their values in a list
}
