% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{process_drug_applications}
\alias{process_drug_applications}
\title{Process Drug Applications}
\usage{
process_drug_applications(data)
}
\arguments{
\item{data}{drug applications to process}
}
\value{
processed drug applications
}
\description{
Process Drug Applications
}
