% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{sourceTypes}
\alias{sourceTypes}
\title{DGIdb Source Types}
\format{
A named list of four character values.
}
\usage{
sourceTypes
}
\value{
A named list of supported DGIdb source-type character values
}
\description{
Supported values for the \code{sourceType} argument to \code{getSources()}.
}
\examples{
sourceTypes
sourceTypes$GENE

}
