% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{get_interactions}
\alias{get_interactions}
\title{Perform an interaction look up for drugs or genes of interest}
\usage{
get_interactions(
  terms,
  use_processing = TRUE,
  search = "genes",
  immunotherapy = NULL,
  antineoplastic = NULL,
  sourcedbname = NULL,
  pmid = NULL,
  interactiontype = NULL,
  approved = NULL
)
}
\arguments{
\item{terms}{drugs or genes for interaction look up}

\item{use_processing}{placeholder}

\item{search}{interaction search type. valid types are "drugs" or "genes"}

\item{immunotherapy}{filter option for results that are used in immunotherapy}

\item{antineoplastic}{filter option for results that are part of antineoplastic regimens # nolint: line_length_linter.}

\item{sourcedbname}{filter option for specific database of interest}

\item{pmid}{filter option for specific PMID}

\item{interactiontype}{filter option for specific interaction types}

\item{approved}{filter option for approved interactions}
}
\value{
interaction results for terms
}
\description{
Perform an interaction look up for drugs or genes of interest
}
\examples{
x <- c("BRAF","PDGFRA")
get_interactions(x)
}
