% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{getAllDrugs}
\alias{getAllDrugs}
\title{Get All Drugs}
\usage{
getAllDrugs(apiUrl = NULL)
}
\arguments{
\item{apiUrl}{DGIdb GraphQL endpoint; defaults to \code{DGIDB_API_URL}.}
}
\value{
A named list containing drug names and concept identifiers.
}
\description{
Gets all drug names and identifiers present in DGIdb.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
getAllDrugs()
\dontshow{\}) # examplesIf}
}
