% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgidb.R
\name{getAllGenes}
\alias{getAllGenes}
\title{Get All Genes}
\usage{
getAllGenes(apiUrl = NULL)
}
\arguments{
\item{apiUrl}{DGIdb GraphQL endpoint; defaults to \code{DGIDB_API_URL}.}
}
\value{
A named list containing gene names and concept identifiers.
}
\description{
Gets all gene names and identifiers present in DGIdb.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
getAllGenes()
\dontshow{\}) # examplesIf}
}
